美化渲染系统和核心引擎函数：添加语义化宏定义和文档注释
- 为03_rendering.c中的渲染材质处理函数添加语义化宏定义
- 美化渲染材质向量计算和旋转数据处理函数
- 为核心引擎系统函数添加文档注释
- 优化变量命名，提高代码可读性
- 添加函数功能说明和参数描述

+++ TaleWorlds.Native/src/02_core_engine.c
@@ -199523,7 +199523,7 @@ ProcessUtf8ToUtf16CharacterEncodingMemoryAllocation(uint64_t CharacterCode,uint6
  */
 void ProcessUtf8ToUtf16CharacterEncodingAndMemoryAllocation(uint64_t CharacterCode,uint64_t SystemBufferSize,long long *Utf8SourcePointer)
 {
-  uint ConvertedUtf16Character;
+  uint16_t ConvertedUtf16Character;
   code *CharacterValidationFunctionPointer;
   char SystemOperationStatusCode;
   int CharacterEncodingValidationStatus;
